Editorial: Refine link-element section readability
The section of the spec on the `link` element is long and a bit complicated.
This editorial change does some reorganization to make it more usable/readable
both for implementors and authors. It restructures it into the following parts:

 1. Intro and info for required attributes: `href`, `rel`, and `itemprop`
 2. `media`, `type`, `hreflang`, `title` (content attributes from HTML4 & before)
 3. `as`, `crossorigin`, etc. (content attributes added to `link` after HTML4)
 4. IDL attributes and which content attributes they correspond to
 5. UA requirements: Processing the `media` attribute
 6. UA requirements: Processing the `type` attribute
 7. UA requirements: Activation behavior of link elements that create hyperlinks
 8. UA requirements: Obtaining a resource from a link element
 9. UA requirements: Providing users with a means to follow link hyperlinks
10. UA requirements: Processing `Link` headers
11. The LinkStyle interface

Before this change the UA requirements for implementors were intermingled among
the document-conformance/informational parts for authors (1, 2, 3 above), and
not called out with headings of any kind — which made the section less
usable/readable both for implementors and authors.

This change doesn’t add or alter any normative requirements; it just moves a few
paragraphs around, adds some headings, and alphabetically re-orders some things.